requirements.txt
python-dotenv
gradio
webvtt-py